Git commit on 19th April

+++ src/main/java/com/email/controller/ApprovedEmailAttachmentController.java
@@ -1,8 +1,11 @@
 package com.email.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.email.entity.ApprovedEmailAttachment;
+import com.email.entity.EmailData;
 import com.email.entity.EmailDataAttachment;
 import com.email.service.ApprovedEmailAttachmentService;
 import com.email.service.EmailDataAttachmentService;
@@ -26,24 +30,43 @@ public class ApprovedEmailAttachmentController {
 	ApprovedEmailAttachmentService service;
 	
 	@PostMapping
-	public ApprovedEmailAttachment create(@RequestBody ApprovedEmailAttachment user){
-	    return service.save(user);
+	public ResponseEntity<ApprovedEmailAttachment> create(@RequestBody ApprovedEmailAttachment email,Principal principal){
+		ApprovedEmailAttachment emailatt=null;
+		if(principal!=null) {
+			emailatt=service.save(email);
+			return new ResponseEntity<ApprovedEmailAttachment>(emailatt,HttpStatus.OK);
+		}
+		return new ResponseEntity<ApprovedEmailAttachment>(emailatt,HttpStatus.UNAUTHORIZED);
+		
 	}
 	
 	@GetMapping
-	public List<ApprovedEmailAttachment> findAll(){
-	  return service.findAll();
+	public ResponseEntity<List<ApprovedEmailAttachment>> findAll(Principal principal){
+		List<ApprovedEmailAttachment> emailList=null;
+		if(principal!=null) {
+			 emailList=service.findAll();
+			return new ResponseEntity<List<ApprovedEmailAttachment>>(emailList,HttpStatus.OK);
+		}
+		return new ResponseEntity<List<ApprovedEmailAttachment>>(emailList,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@PutMapping(value="/{id}")
-	public ApprovedEmailAttachment updateById(@PathVariable Long id, @RequestBody ApprovedEmailAttachment user) {
-		return service.save(user);
+	public ResponseEntity<ApprovedEmailAttachment> updateById(@PathVariable Long id, @RequestBody ApprovedEmailAttachment email,Principal principal) {
+		ApprovedEmailAttachment emailatt=null;
+		if(principal!=null) {
+			 emailatt=service.save(email);
+			return new ResponseEntity<ApprovedEmailAttachment>(emailatt,HttpStatus.OK);
+		}
+		return new ResponseEntity<ApprovedEmailAttachment>(emailatt,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@DeleteMapping(path ={"/{id}"})
-	public String deleteById(@PathVariable Long id) {
-		service.deleteById(id);
-		return "success..";
+	public ResponseEntity<String> deleteById(@PathVariable Long id,Principal principal) {
+		if(principal!=null) {
+			service.deleteById(id);
+			return new ResponseEntity<String>("success..",HttpStatus.OK);
+		}
+		return new ResponseEntity<String>("Failed..",HttpStatus.UNAUTHORIZED);
 	}
