+++ src/main/java/com/email/controller/EmailDataAttachmentController.java
@@ -1,8 +1,11 @@
 package com.email.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.email.entity.ApprovedEmail;
 import com.email.entity.EmailDataAttachment;
 import com.email.service.EmailDataAttachmentService;
 
@@ -22,24 +26,43 @@ public class EmailDataAttachmentController {
 	EmailDataAttachmentService service;
 	
 	@PostMapping
-	public EmailDataAttachment create(@RequestBody EmailDataAttachment user){
-	    return service.save(user);
+	public ResponseEntity<EmailDataAttachment> create(@RequestBody EmailDataAttachment user,Principal principal){
+		EmailDataAttachment appEmail=null;
+		if(principal!=null) {
+			appEmail=service.save(user);
+			return new ResponseEntity<EmailDataAttachment>(appEmail,HttpStatus.OK);
+		}
+		return new ResponseEntity<EmailDataAttachment>(appEmail,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@GetMapping
-	public List<EmailDataAttachment> findAll(){
-	  return service.findAll();
+	public ResponseEntity<List<EmailDataAttachment>> findAll(Principal principal){
+		
+		List<EmailDataAttachment> appEmail=null;
+		if(principal!=null) {
+			appEmail= service.findAll();
+			return new ResponseEntity<List<EmailDataAttachment>>(appEmail,HttpStatus.OK);
+		}
+		return new ResponseEntity<List<EmailDataAttachment>>(appEmail,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@PutMapping(value="/{id}")
-	public EmailDataAttachment updateById(@PathVariable Long id, @RequestBody EmailDataAttachment user) {
-		return service.save(user);
+	public ResponseEntity<EmailDataAttachment> updateById(@PathVariable Long id, @RequestBody EmailDataAttachment user,Principal principal) {
+		EmailDataAttachment appEmail=null;
+		if(principal!=null) {
+			appEmail=service.save(user);
+			return new ResponseEntity<EmailDataAttachment>(appEmail,HttpStatus.OK);
+		}
+		return new ResponseEntity<EmailDataAttachment>(appEmail,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@DeleteMapping(path ={"/{id}"})
-	public String deleteById(@PathVariable Long id) {
-		service.deleteById(id);
-		return "success..";
+	public ResponseEntity<String> deleteById(@PathVariable Long id,Principal principal) {
+		if(principal!=null) {
+			service.deleteById(id);
+			return new ResponseEntity<String>("success..",HttpStatus.OK);
+		}
+		return new ResponseEntity<String>("Failed..",HttpStatus.UNAUTHORIZED);
 	}
