+++ src/main/java/com/email/controller/SpecificationController.java
@@ -1,8 +1,11 @@
 package com.email.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,24 +28,44 @@ public class SpecificationController {
 	SpecificationService service;
 	
 	@PostMapping
-	public Specification create(@RequestBody Specification user){
-	    return service.save(user);
+	public ResponseEntity<Specification> create(@RequestBody Specification user,Principal principal){
+		Specification appEmail=null;
+		if(principal!=null) {
+			appEmail=service.save(user);
+			return new ResponseEntity<Specification>(appEmail,HttpStatus.OK);
+		}
+		return new ResponseEntity<Specification>(appEmail,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@GetMapping
-	public List<Specification> findAll(){
-	  return service.findAll();
+	public ResponseEntity<List<Specification>> findAll(Principal principal){
+		List<Specification> appEmail=null;
+		if(principal!=null) {
+			appEmail= service.findAll();
+			return new ResponseEntity<List<Specification>>(appEmail,HttpStatus.OK);
+		}
+		return new ResponseEntity<List<Specification>>(appEmail,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@PutMapping(value="/{id}")
-	public Specification updateById(@PathVariable Long id, @RequestBody Specification user) {
-		return service.save(user);
+	public ResponseEntity<Specification> updateById(@PathVariable Long id, @RequestBody Specification user,Principal principal) {
+		Specification appEmail=null;
+		if(principal!=null) {
+			appEmail=service.save(user);
+			return new ResponseEntity<Specification>(appEmail,HttpStatus.OK);
+		}
+		return new ResponseEntity<Specification>(appEmail,HttpStatus.UNAUTHORIZED);
 	}
 	
 	@DeleteMapping(path ={"/{id}"})
-	public String deleteById(@PathVariable Long id) {
-		service.deleteById(id);
-		return "success..";
+	public ResponseEntity<String> deleteById(@PathVariable Long id,Principal principal) {
+		
+		if(principal!=null) {
+			service.deleteById(id);
+			return new ResponseEntity<String>("success..",HttpStatus.OK);
+		}
+		return new ResponseEntity<String>("Failed..",HttpStatus.UNAUTHORIZED);
+		
 	}
